pyproject.toml
[build-system]
requires = [
  "wheel",
  "setuptools"
]
build-backend = "setuptools.build_meta"

[project]
name = "rAIght.move"
version = "0.1"
license = {file = "LICENSE"}
description = "App that helps with your exercise performace"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "matplotlib>=3.8.2",
  "mediapipe==0.10.9",
  "numpy>=1.26.2",
  "opencv-python>=4.8.1.78",
  "pandas>=2.1.4",
  "setuptools>=65.5.0",
  "dtw-python>=1.3.1",
  "pyyaml>=6.0.1",
  "scikit-learn>=1.4.0"
]

[project.optional-dependencies]
dev = [
  "isort>=5.13.2",
  "black>=23.12.0"
]

[tool.setuptools]
package-dir = {"" = "raight_move"}
